+++ SmartTalent.Domain/Models/RoomType.cs
@@ -19,6 +19,6 @@ namespace SmartTalent.Domain.Models
         [StringLength(20)]
         public string Name { get; set; }
 
-        public int ValuePerNight { get; set; }
+        public double ValuePerNight { get; set; }
     }
 }
